remove non-needed code
