--- apps/backoffice/src/app/(authenticated)/profile/components/user-info/user-info.module.scss
@@ -1,10 +1,0 @@
-.icon {
-    color: light-dark(var(--mantine-color-gray-5), var(--mantine-color-dark-3));
-  }
-  
-  .name {
-    font-family:
-      Greycliff CF,
-      var(--mantine-font-family),
-      sans-serif;
-  }
