--- apps/backoffice/src/app/(authenticated)/profile/layout.tsx
@@ -1,15 +1,0 @@
-import React, { Suspense } from 'react';
-import { Title, Text } from '@mantine/core';
-import { useTranslations } from 'next-intl';
-import LoaderCentered from '@/components/misc/loader-centered';
-
-export default function RootLayout({ children }: { children: React.ReactNode }): React.ReactNode {
-  const t = useTranslations('profile');
-  return (
-    <>
-      <Title mb="md">{t('title')}</Title>
-      <Text mb="xl">{t('description')}</Text>
-      <Suspense fallback={<LoaderCentered />}>{children}</Suspense>
-    </>
-  );
-}
