+++ apps/backoffice/src/app/layout.tsx
@@ -8,7 +8,10 @@ import { NextIntlClientProvider } from 'next-intl';
 import { Analytics } from '@vercel/analytics/react';
 import { SpeedInsights } from '@vercel/speed-insights/next';
 import { getLocale, getMessages } from 'next-intl/server';
-import NotificationsHandler from '@/components/misc/notifications-handler';
+import { Suspense } from 'react';
+import { MainAppShell } from '@/components/shells/main-shell/main-shell';
+import SetUserSentry from '@/app/set-user-sentry';
+import LoaderCentered from '@/components/misc/loader-centered';
 
 export const metadata: Metadata = {
   title: 'AdsViewer',
@@ -25,8 +28,10 @@ export default async function RootLayout({ children }: { children: React.ReactNo
         <ColorSchemeScript />
         <NextIntlClientProvider messages={messages}>
           <MantineProvider defaultColorScheme="auto">
-            <NotificationsHandler />
-            {children}
+            <MainAppShell>
+              <SetUserSentry />
+              <Suspense fallback={<LoaderCentered />}>{children}</Suspense>
+            </MainAppShell>
             <Analytics />
             <SpeedInsights />
           </MantineProvider>
