+++ apps/backoffice/src/components/user-button/user-button.tsx
@@ -1,22 +1,20 @@
 'use client';
 
-import { Group, Avatar, Text, Flex } from '@mantine/core';
-import { useCallback, useEffect, useState, type ReactNode } from 'react';
-import { useRouter } from 'next/navigation';
+import { Avatar, Flex, Group, Text } from '@mantine/core';
+import { type ReactNode, useCallback, useEffect, useState } from 'react';
 import { useAtom } from 'jotai';
 import { logger } from '@repo/logger';
 import { notifications } from '@mantine/notifications';
 import { useTranslations } from 'next-intl';
-import { getUserDetails } from '@/app/(authenticated)/actions';
+import { getUserDetails } from '@/app/actions';
 import LoaderCentered from '@/components/misc/loader-centered';
 import { userDetailsAtom } from '@/app/atoms/user-atoms';
-import { IntegrationStatus, type Integration } from '@/graphql/generated/schema-server';
+import { type Integration, IntegrationStatus } from '@/graphql/generated/schema-server';
 import classes from './user-button.module.scss';
 
 export default function UserButton(): ReactNode {
   const tGeneric = useTranslations('generic');
   const tIntegrations = useTranslations('integrations');
-  const router = useRouter();
   const [userDetails, setUserDetails] = useAtom(userDetailsAtom);
   const [isDataLoaded, setIsDataLoaded] = useState<boolean>(false);
 
@@ -49,14 +47,10 @@ export default function UserButton(): ReactNode {
       });
   }, [checkIntegrationTokensForExpiration, setUserDetails]);
 
-  const redirectToProfile = (): void => {
-    router.push('profile');
-  };
-
   return (
     <Flex justify="flex-start" className={classes.user}>
       {isDataLoaded ? (
-        <Group onClick={redirectToProfile}>
+        <Group>
           <Avatar src={userDetails.photoUrl} radius="xl" />
 
           <div style={{ flex: 1 }}>
