+++ apps/backoffice/src/env.mjs
@@ -10,9 +10,7 @@ export const env = createEnv({
    *  - env | grep -e EXAMPLE_1 -e EXAMPLE_2 >> apps/backoffice/.env.production
    */
   server: {
-    NODE_ENV: z.enum(['development', 'test', 'production']),
     AUTH_SECRET: z.string().min(1).default('something'),
-    REFRESH_SECRET: z.string().min(1).default('refreshSecret'),
     BACKOFFICE_ENDPOINT: z.string().min(1).default('http://localhost:3002'),
   },
 
@@ -25,7 +23,6 @@ export const env = createEnv({
     // NEXT_PUBLIC_CLIENTVAR: z.string().min(1),
     // Add `.min(1) on these if you want to make sure they're not empty
     NEXT_WEBAPP_ENDPOINT: z.string().min(1).default('http://localhost:3000'),
-    NEXT_PUBLIC_GRAPHQL_ENDPOINT: z.string().min(1).default('http://localhost:3002/graphql'),
     NEXT_PUBLIC_REAL_GRAPHQL_ENDPOINT: z.string().min(1).default('http://localhost:4000/graphql'),
   },
 
@@ -34,11 +31,9 @@ export const env = createEnv({
    * middlewares) or client-side so we need to destruct manually.
    */
   runtimeEnv: {
-    NODE_ENV: process.env.NODE_ENV,
     AUTH_SECRET: process.env.AUTH_SECRET,
     BACKOFFICE_ENDPOINT: process.env.BACKOFFICE_ENDPOINT,
     NEXT_WEBAPP_ENDPOINT: process.env.NEXT_WEBAPP_ENDPOINT,
-    NEXT_PUBLIC_GRAPHQL_ENDPOINT: process.env.NEXT_PUBLIC_GRAPHQL_ENDPOINT,
     NEXT_PUBLIC_REAL_GRAPHQL_ENDPOINT: process.env.NEXT_PUBLIC_REAL_GRAPHQL_ENDPOINT,
   },
   /**
